requirements.txt
torch
lightning
scikit-learn
pandas
mlflow
torchinfo
country_converter
pyyaml
